requirements.txt
appdirs==1.4.3
bcrypt==3.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
ciscoconfparse==1.5.19
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.1
distlib==0.3.0
distro==1.4.0
dnspython==2.0.0
future==0.18.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipaddress==1.0.23
Jinja2==2.11.2
junos-eznc==2.5.3
lockfile==0.12.2
lxml==4.5.2
MarkupSafe==1.1.1
msgpack==0.6.2
mypy-extensions==0.4.3
napalm==3.2.0
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.3.0
nornir==3.0.0b1
nornir-napalm==0.1.1
nornir-netbox==0.1.1
nornir-utils==0.1.0
ntc-templates==1.5.0
packaging==20.3
paramiko==2.7.2
passlib==1.7.2
pep517==0.8.2
progress==1.5
pycparser==2.20
pyeapi==0.8.3
PyNaCl==1.4.0
pyparsing==2.4.6
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.14.0
python-netbox==0.0.16
pytoml==0.1.21
PyYAML==5.3.1
requests==2.24.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scp==0.13.2
six==1.14.0
textfsm==1.1.0
transitions==0.8.2
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
yamlordereddictloader==0.4.0
